requirements.txt
# Core dependencies
httpx>=0.25.0
python-dotenv>=1.0.0
websocket-client>=1.6.0

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
ruff>=0.1.0
